+++ CHANGES.md
@@ -9,6 +9,7 @@ Beta 6 (ongoing)
 * added stop() support to $.request()
 * added parentNum parameter to up()
 * values() uses ids if a form element has no name
+* support for bubble selectors in onChange(), onOver(), onFocus() and onClick()
 * changed signatures of request's promise handlers: the success handler is called with the XHR object instead of 
   'responseXml' property as second argument. The failure handler has no 'statusText' anymore (who needs that?).
   Instead it gets the XHR object as last argument.